changed naming, added async await middleware
