+++ src/middleware/asyncMiddleware.js
@@ -1,0 +1,6 @@
+export default function asyncMiddleware(func) {
+  return (req, res, next) =>
+    Promise
+    .resolve(func(req, res, next))
+    .catch(next);
+}
